Use DelegateTo to delegate chain components
