--- crates/chain/chain-components/src/impls/delegate/message_builders/mod.rs
@@ -1,4 +1,0 @@
-pub mod channel_handshake;
-pub mod connection_handshake;
-pub mod create_client;
-pub mod update_client;
