--- crates/chain/chain-components/src/impls/delegate/message_builders/update_client.rs
@@ -1,27 +1,0 @@
-use alloc::vec::Vec;
-use core::marker::PhantomData;
-
-use cgp::prelude::{DelegateComponent, HasErrorType};
-
-use crate::traits::message_builders::update_client::UpdateClientMessageBuilder;
-use crate::traits::types::ibc::HasIbcChainTypes;
-use crate::traits::types::update_client::HasUpdateClientPayloadType;
-
-pub struct DelegateBuildUpdateClientMessage<Components>(pub PhantomData<Components>);
-
-impl<Chain, Counterparty, Components, Delegate> UpdateClientMessageBuilder<Chain, Counterparty>
-    for DelegateBuildUpdateClientMessage<Components>
-where
-    Chain: HasIbcChainTypes<Counterparty> + HasErrorType,
-    Counterparty: HasUpdateClientPayloadType<Chain>,
-    Delegate: UpdateClientMessageBuilder<Chain, Counterparty>,
-    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
-{
-    async fn build_update_client_message(
-        chain: &Chain,
-        client_id: &Chain::ClientId,
-        payload: Counterparty::UpdateClientPayload,
-    ) -> Result<Vec<Chain::Message>, Chain::Error> {
-        Delegate::build_update_client_message(chain, client_id, payload).await
-    }
-}
