--- crates/chain/chain-components/src/impls/delegate/mod.rs
@@ -1,3 +1,0 @@
-pub mod message_builders;
-pub mod message_height;
-pub mod queries;
