--- crates/chain/chain-components/src/impls/delegate/queries/mod.rs
@@ -1,3 +1,0 @@
-pub mod client_state;
-pub mod consensus_state;
-pub mod consensus_state_heights;
