+++ crates/chain/chain-components/src/traits/message_builders/create_client.rs
@@ -1,3 +1,4 @@
+use cgp::core::component::DelegateTo;
 use cgp::prelude::*;
 
 use crate::traits::types::create_client::{
@@ -17,4 +18,22 @@ where
         create_client_options: &Self::CreateClientMessageOptions,
         counterparty_payload: Counterparty::CreateClientPayload,
     ) -> Result<Self::Message, Self::Error>;
+}
+
+impl<Chain, Counterparty, Components, Delegate> CreateClientMessageBuilder<Chain, Counterparty>
+    for DelegateTo<Components>
+where
+    Chain: HasCreateClientMessageOptionsType<Counterparty> + HasMessageType + HasErrorType,
+    Counterparty: HasCreateClientPayloadType<Chain>,
+    Delegate: CreateClientMessageBuilder<Chain, Counterparty>,
+    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
+{
+    async fn build_create_client_message(
+        chain: &Chain,
+        create_client_options: &Chain::CreateClientMessageOptions,
+        counterparty_payload: Counterparty::CreateClientPayload,
+    ) -> Result<Chain::Message, Chain::Error> {
+        Delegate::build_create_client_message(chain, create_client_options, counterparty_payload)
+            .await
+    }
 }
