+++ crates/chain/chain-components/src/traits/message_builders/update_client.rs
@@ -1,5 +1,6 @@
 use alloc::vec::Vec;
 
+use cgp::core::component::DelegateTo;
 use cgp::prelude::*;
 
 use crate::traits::types::ibc::HasIbcChainTypes;
@@ -17,4 +18,21 @@ where
         client_id: &Self::ClientId,
         payload: Counterparty::UpdateClientPayload,
     ) -> Result<Vec<Self::Message>, Self::Error>;
+}
+
+impl<Chain, Counterparty, Components, Delegate> UpdateClientMessageBuilder<Chain, Counterparty>
+    for DelegateTo<Components>
+where
+    Chain: HasIbcChainTypes<Counterparty> + HasErrorType,
+    Counterparty: HasUpdateClientPayloadType<Chain>,
+    Delegate: UpdateClientMessageBuilder<Chain, Counterparty>,
+    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
+{
+    async fn build_update_client_message(
+        chain: &Chain,
+        client_id: &Chain::ClientId,
+        payload: Counterparty::UpdateClientPayload,
+    ) -> Result<Vec<Chain::Message>, Chain::Error> {
+        Delegate::build_update_client_message(chain, client_id, payload).await
+    }
 }
