+++ crates/chain/chain-components/src/traits/queries/client_state.rs
@@ -1,5 +1,6 @@
 use alloc::vec::Vec;
 
+use cgp::core::component::DelegateTo;
 use cgp::prelude::*;
 
 use crate::traits::queries::chain_status::CanQueryChainStatus;
@@ -133,5 +134,55 @@ where
 
         self.query_all_client_states(Chain::chain_status_height(&status))
             .await
+    }
+}
+
+impl<Chain, Counterparty, Components, Delegate> ClientStateQuerier<Chain, Counterparty>
+    for DelegateTo<Components>
+where
+    Chain: HasIbcChainTypes<Counterparty> + HasErrorType,
+    Counterparty: HasClientStateType<Chain>,
+    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
+    Delegate: ClientStateQuerier<Chain, Counterparty>,
+{
+    async fn query_client_state(
+        chain: &Chain,
+        client_id: &Chain::ClientId,
+        height: &Chain::Height,
+    ) -> Result<Counterparty::ClientState, Chain::Error> {
+        Delegate::query_client_state(chain, client_id, height).await
+    }
+}
+
+impl<Chain, Counterparty, Components, Delegate> ClientStateWithProofsQuerier<Chain, Counterparty>
+    for DelegateTo<Components>
+where
+    Chain: HasIbcChainTypes<Counterparty> + HasCommitmentProofType + HasErrorType,
+    Counterparty: HasClientStateType<Chain>,
+    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
+    Delegate: ClientStateWithProofsQuerier<Chain, Counterparty>,
+{
+    async fn query_client_state_with_proofs(
+        chain: &Chain,
+        client_id: &Chain::ClientId,
+        height: &Chain::Height,
+    ) -> Result<(Counterparty::ClientState, Chain::CommitmentProof), Chain::Error> {
+        Delegate::query_client_state_with_proofs(chain, client_id, height).await
+    }
+}
+
+impl<Chain, Counterparty, Components, Delegate> AllClientStatesQuerier<Chain, Counterparty>
+    for DelegateTo<Components>
+where
+    Chain: HasIbcChainTypes<Counterparty> + HasErrorType,
+    Counterparty: HasClientStateType<Chain>,
+    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
+    Delegate: AllClientStatesQuerier<Chain, Counterparty>,
+{
+    async fn query_all_client_states(
+        chain: &Chain,
+        height: &Chain::Height,
+    ) -> Result<Vec<(Chain::ClientId, Counterparty::ClientState)>, Chain::Error> {
+        Delegate::query_all_client_states(chain, height).await
     }
 }
