+++ crates/chain/chain-components/src/traits/queries/consensus_state.rs
@@ -1,3 +1,4 @@
+use cgp::core::component::DelegateTo;
 use cgp::prelude::*;
 
 use super::chain_status::CanQueryChainStatus;
@@ -95,6 +96,48 @@ where
             client_id,
             consensus_height,
             Chain::chain_status_height(&status),
+        )
+        .await
+    }
+}
+
+impl<Chain, Counterparty, Components, Delegate> ConsensusStateQuerier<Chain, Counterparty>
+    for DelegateTo<Components>
+where
+    Chain: HasIbcChainTypes<Counterparty> + HasErrorType,
+    Counterparty: HasConsensusStateType<Chain> + HasHeightType,
+    Delegate: ConsensusStateQuerier<Chain, Counterparty>,
+    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
+{
+    async fn query_consensus_state(
+        chain: &Chain,
+        client_id: &Chain::ClientId,
+        consensus_height: &Counterparty::Height,
+        query_height: &Chain::Height,
+    ) -> Result<Counterparty::ConsensusState, Chain::Error> {
+        Delegate::query_consensus_state(chain, client_id, consensus_height, query_height).await
+    }
+}
+
+impl<Chain, Counterparty, Components, Delegate> ConsensusStateWithProofsQuerier<Chain, Counterparty>
+    for DelegateTo<Components>
+where
+    Chain: HasIbcChainTypes<Counterparty> + HasCommitmentProofType + HasErrorType,
+    Counterparty: HasConsensusStateType<Chain> + HasHeightType,
+    Delegate: ConsensusStateWithProofsQuerier<Chain, Counterparty>,
+    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
+{
+    async fn query_consensus_state_with_proofs(
+        chain: &Chain,
+        client_id: &Chain::ClientId,
+        consensus_height: &Counterparty::Height,
+        query_height: &Chain::Height,
+    ) -> Result<(Counterparty::ConsensusState, Chain::CommitmentProof), Chain::Error> {
+        Delegate::query_consensus_state_with_proofs(
+            chain,
+            client_id,
+            consensus_height,
+            query_height,
         )
         .await
     }
